requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.2
pyocr==0.4.2
python-Levenshtein==0.12.0
six==1.10.0
Werkzeug==0.11.11
